fix compile error with clang
* random.c (fill_random_bytes_syscall): fix compile error with
  clang. [ruby-core:69931] [Bug #11343]

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@51240 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ random.c
@@ -521,7 +521,7 @@ fill_random_bytes_syscall(void *seed, size_t size)
 {
     static rb_atomic_t try_syscall = 1;
     if (try_syscall) {
-	int ret;
+	long ret;
 	errno = 0;
 	ret = syscall(SYS_getrandom, seed, size, 0);
 	if (errno == ENOSYS) {
